brain_tumor_detection/requirements.txt
streamlit
joblib
scikit-learn
numpy
pillow
